+++ app/models/randomization_algorithm/algorithms/minimization.rb
@@ -116,27 +116,9 @@ module RandomizationAlgorithm
               past_distributions: past_distributions,
               weighted_eligible_arms: weighted_eligible_arms.collect{ |arm| { name: arm.name, id: arm.id } }.sort{|a,b| a[:name] <=> b[:name]}
             )
-            @randomization_scheme.stratification_factors.each do |sf|
-              if criteria = criteria_pairs.select{|sfid, oid| sfid == sf.id}.first
-                if sf.stratifies_by_site?
-                  randomization.randomization_characteristics.create(
-                    project_id: @randomization_scheme.project_id,
-                    randomization_scheme_id: @randomization_scheme.id,
-                    list_id: randomization.list_id,
-                    stratification_factor_id: sf.id,
-                    site_id: @randomization_scheme.project.sites.where(id: criteria.last).pluck(:id).first
-                  )
-                else
-                  randomization.randomization_characteristics.create(
-                    project_id: @randomization_scheme.project_id,
-                    randomization_scheme_id: @randomization_scheme.id,
-                    list_id: randomization.list_id,
-                    stratification_factor_id: sf.id,
-                    stratification_factor_option_id: sf.stratification_factor_options.where(id: criteria.last).pluck(:id).first
-                  )
-                end
-              end
-            end
+
+            # Add Characteristics
+            self.add_randomization_characteristics!(randomization, criteria_pairs)
           end
 
           randomization