+++ CHANGELOG.md
@@ -5,6 +5,7 @@
   - Project owners and editors can now add one or more randomization schemes to a project
   - Treatment Arm distributions by stratification factors are shown on randomization scheme page
   - Two types of randomization algorithms are available to use
+  - Randomization criteria are now stored when subject is randomized
   - **Permuted-block Algorithm**
     - Treatment Arms can now be specified along with a weight allocation
     - Different block sizes and allocations can now be specified